Moved utility dataflow functions (semijoin_arranged, antijoin_arranged, concatenate_collections) to program/arrange.rs
